% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{get_njtr1}
\alias{get_njtr1}
\title{Download NJ car accident records for a given year}
\usage{
get_njtr1(year, type)
}
\arguments{
\item{year}{Year to download crash data}

\item{type}{The type of NJDOT crash data to download.}
}
\value{
tibble of all reported accidents for the year
}
\description{
Download NJ car accident records for a given year
}
\examples{
get_njtr1(year = 2019, type = 'Accidents')
}
