% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{read_njtr1}
\alias{read_njtr1}
\title{Read an NJTR-1 data file that was previously downloaded}
\usage{
read_njtr1(file, geo = FALSE)
}
\arguments{
\item{file}{path to a file downloaded from NJDOT}

\item{geo}{Logical, whether to filter only to geotagged cases (default = FALSE)}
}
\value{
tibble of cleaned & parsed NJTR-1 data
}
\description{
Read an NJTR-1 data file that was previously downloaded
}
\examples{
read_njtr1(system.file("extdata/testdata/Ocean2019Accidents.zip", package = "njtr1"))
}
